# docstrings, process xor, rotate optimisations