Analog implementation to Python PR. 

Benchmarks for loops:
```csharp
using System;
using System.Diagnostics;
using System.Runtime;

namespace demo
{
    class MainClass
    {
        static bool ByForArrayLength (byte[] data)
        {
            for (int i = 0; i < data.Length; i++)
                if (data [i] != 0)
                    return false;
            return true;
        }

        static bool ByForLocalLength (byte[] data)
        {
            int len = data.Length;
            for (int i = 0; i < len; i++)
                if (data [i] != 0)
                    return false;
            return true;
        }

        static bool ByForeach (byte[] data)
        {
            foreach (byte b in data)
                if (b != 0)
                    return false;
            return true;
        }

        static void Measure (Action work, string description)
        {
            var watch = Stopwatch.StartNew ();
            work.Invoke ();
            Console.WriteLine ("{0,-40}: {1} ms", description, watch.Elapsed.TotalMilliseconds);
        }

        public static void Main (string[] args)
        {
            GCSettings.LatencyMode = GCLatencyMode.LowLatency;
            byte[] data = new byte[256 * 1024 * 1024];

            Measure (() => ByForArrayLength (data), "For with byte[].Length property");
            Measure (() => ByForLocalLength (data), "For with Length in local variable");
            Measure (() => ByForeach (data), "Foreach over byte[]");
        }
    }
}
```
```
$ mcs Program.cs -optimize
$ ./Program.exe 
For with byte[].Length property         : 427,2671 ms
For with Length in local variable       : 331,5512 ms
Foreach over byte[]                     : 272,0288 ms
```
Benchmarks for rotations:
```csharp
using System;
using System.Diagnostics;
using System.Runtime;

namespace demo
{
    class MainClass
    {
        static void ByShifts (byte[] data)
        {
            int len = data.Length;
            int amount = 1;
            int anti = 7;

            for (int i = 0; i < len; i++)
            {
                byte x = data[i];
                data[i] = (byte)((x << amount) | (x >> anti));
            }
        }

        static void ByLookup (byte[] data)
        {
            int len = data.Length;
            byte[] translate = new byte[256];

            for (int i = 0; i < len; i++)
                data[i] = translate[data[i]];
        }

        static void Measure (Action work, string description)
        {
            var watch = Stopwatch.StartNew ();
            work.Invoke ();
            Console.WriteLine ("{0,-40}: {1} ms", description, watch.Elapsed.TotalMilliseconds);
        }

        public static void Main (string[] args)
        {
            GCSettings.LatencyMode = GCLatencyMode.LowLatency;
            byte[] data = new byte[256 * 1024 * 1024];

            Measure (() => ByShifts (data), "Loop with << >> |");
            Measure (() => ByLookup (data), "Loop with translate[...]");
        }
    }
}
```
```
$ mcs Program.cs -optimize
$ ./Program.exe 
Loop with << >> |                       : 808,9077 ms
Loop with translate[...]                : 513,9755 ms
```
Requesting review also from @Arlorean and @Kolanich .