@arekbulski 2018-04-06
Check last commit, fixed all suggestions.