@arekbulski 2018-04-06
Fixed all outstanding suggestinons, check last (amended) commit.