@arekbulski 2018-04-06
I appologise for putting so much stuff into a single PR, but I could not really separate the docstrings from the implementation, because XOR ROL implementations also needed to change the docstrings. Having 2 PRs where one superces another would not allow me to rebase it if needed. Please consider this PR frozen at this point, unless @GreyCat makes any additional suggestions.