@arekbulski 2018-04-06
There are still 2 methods with ??? docstrings.