@arekbulski 2018-04-06
And I just said it was RTM... well, it is. It was. And yet I still found _1 more line to update_...
Its in last separate commit, not amended.