@arekbulski 2018-04-06
Found _yet another_ small optimisation: Stripping and terminating bytes can have an edge case where entire data would be returned, in which case, it returns original array instead of copying it.
Its in last separate commit, not amended.