@arekbulski 2018-04-11
Saving currect hashid for undo: 52308c2ca00c50d7dabbca47fac5e94ed4f840b9