@arekbulski 2018-04-11
All commits were amended to reflect last benchmarks. Previous changset was 52308c2ca00c50d7dabbca47fac5e94ed4f840b9.