@Arlorean 2018-04-16
My experience in real-world .NET applications on Windows is that the amount of memory you turn over, which puts pressure on the Garbage Collector, often outweighs any micro-optimizations at the loop level like this. We use tools like [dotTrace ](https://www.jetbrains.com/profiler/) and [dotMemory](https://www.jetbrains.com/dotmemory/) to pin down performance problems, but more often than not, it's a change in algorithm, rather than a change in loop structuring that provides the most performance gains.

The Roslyn team have [some guidelines](https://github.com/dotnet/roslyn/wiki/Contributing-Code#coding-conventions) which may help. The key one being _don't use LINQ_ as it allocates memory which causes performance issues.

This brings me to an overall concern about the KaitaiStruct C# API. As much as I love it, I don't love the amount of memory that gets allocated when parsing a file. A solution along the lines of [FlatBuffers](https://google.github.io/flatbuffers/) would be preferable since it performs **zero memory allocations** but still gives easy access to the data structures stored in essentially a byte array.