@GreyCat 2018-04-17
@arekbulski 
> I looked at the code you quoted for few days, what is so wrong with it?

Ok, the catch is that you've used Python formatting, missing `{}`s, and this is #. Formatted properly, your code does the following:

```csharp
int k = 0;
for (int i = 0; i < data.Length; i++)
{
    data[i] = (byte) (data[i] ^ key[k]);
}
k++;
if (k == key.Length)
{
    k = 0;
}
```

That `k` wrapped increment fell out of the loop.