@arekbulski 2018-04-17
Ha ha, that is some sloppy code. Oh man. I am busy with another project, so I will fix it but not right away. PR ON HOLD.